\name{make.family.rsm}
\alias{make.family.rsm}
\title{
  Support for RSM Family Functions
}
\description{
  This is support for the functions \code{\link{family.rsm}}, 
  \code{\link{extreme}}, \code{\link{Huber}}, \code{\link{logistic}},  
  \code{\link{logWeibull}} and \code{\link{student}}.  It is not intended to 
  be called directly by users.
}
\usage{
make.family.rsm(name, arg, \dots)
}
\keyword{internal}
\keyword{programming}

