\name{rsm.dispersion}
\alias{rsm.dispersion}
\title{
  Support for Functions rsm.fit and rsm.surv
}
\description{
  This is support for the functions \code{\link{rsm.fit}} and 
  \code{\link{rsm.surv}}.  It is not intended to be called directly by 
  users.
}
\usage{
rsm.dispersion(log.dispersion, abs.res, family, arg)
}
\keyword{internal}
\keyword{regression}
\keyword{programming}
